Fix build errors for Netlify deployment
- Replaced broken AuthContext.tsx with working AuthContextWithMock
- Removed async rewrites from next.config.js (not compatible with static export)
- Fixed syntax errors preventing build
